UI nodes observe on world loading, and fix window closing behavior
