+++ Ciallo/NodeControl/LayerPanel.cs
@@ -3,6 +3,7 @@ using System;
 using Arch.Core;
 using Arch.Core.Extensions;
 using Ciallo.Data;
+using Ciallo.Misc;
 using R3;
 
 namespace Ciallo.NodeControl;
@@ -12,40 +13,23 @@ namespace Ciallo.NodeControl;
 /// </summary>
 public partial class LayerPanel : VBoxContainer
 {
-    private Control _visibleLayerTree;
-    
     public override void _Ready()
     {
         GetNode<Node>("%LayerContainerPreview").QueueFree();
-        AppWorldManager.WorkingWorld.Skip(1).Subscribe(w =>
-        {
-            if(w == null && _visibleLayerTree != null)
-            {
-                _visibleLayerTree.Visible = false;
-                _visibleLayerTree = null;
-                return;
-            }
-            if (_visibleLayerTree != null) _visibleLayerTree.Visible = false;
-            var doc = w.Document();
-            _visibleLayerTree = doc.Get<LayerContainer>();
-            _visibleLayerTree.Visible = true;
-        }).AddTo(this);
     }
 
     public void CreateAddLayerContainer(Entity document)
     {
         var layerContainer = LayerContainer.Instantiate();
-        layerContainer.Visible = false;
+        layerContainer.VisibleIf(AppWorldManager.WorkingDocument, document);
         AddChild(layerContainer);
         document.Add(layerContainer);
     }
 
     public void RemoveFreeLayerContainer(Entity document)
     {
-        var layerTreeControl = document.Get<LayerContainer>();
-        if(_visibleLayerTree == layerTreeControl)
-            _visibleLayerTree = null;
+        var layerContainer = document.Get<LayerContainer>();
         document.Remove<LayerContainer>();
-        layerTreeControl.QueueFree();
+        layerContainer.QueueFree();
     }
 }
